streamlit/requirements.txt
beautifulsoup4==4.13.4
matplotlib==3.10.5
matplotlib-inline==0.1.7
mysql-connector-python==9.4.0
numpy==2.3.2
pandas==2.3.2
python-dotenv==1.1.1
requests==2.32.4
selenium==4.35.0
streamlit==1.48.1